refactor: relocation-service로 네이밍 변경
