+++ AGENTS.md
@@ -64,7 +64,7 @@
 - 하나의 workflow 안에서 변경 감지 기반 모듈별 빌드를 수행한다.
 - 루트 Gradle 설정, Gradle Wrapper, CI workflow가 바뀌면 3개 서비스 빌드를 모두 수행한다.
 - `common-core`, `common-web`이 바뀌면 3개 서비스 빌드를 모두 수행한다.
-- `common-kakao`가 바뀌면 `dispatch-service`, `redispatch-service` 빌드를 수행한다.
+- `common-kakao`가 바뀌면 `dispatch-service`, `relocation-service` 빌드를 수행한다.
 - 서비스 모듈만 바뀌면 해당 서비스 빌드만 수행한다.
 - 각 서비스 빌드는 `./gradlew :{service}:clean :{service}:build` 형태로 실행한다.
 - CD 단계는 추가하지 않는다.
