+++ relocation-service/src/test/kotlin/com/baro/relocation/RelocationServiceApplicationTests.kt
@@ -1,10 +1,10 @@
-package com.baro.redispatch
+package com.baro.relocation
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
 
 @SpringBootTest
-class RedispatchServiceApplicationTests {
+class RelocationServiceApplicationTests {
     @Test
     fun `스프링 컨텍스트가 정상적으로 로드된다`() {
     }
